refactor & add simplify mul

+++ Input.h
@@ -10,7 +10,7 @@ ErrorCode tree_get_val_vars (Tree* tree);
 ErrorCode tree_get_num_var  (Tree* tree, int* num_var);
 ErrorCode tree_read         (Tree* tree, char** buf);
 ErrorCode file_to_buf       (const char* name_file, char** buf);
-ErrorCode get_var           (Variables* vars, int number_var);
+ErrorCode get_var           (Variables* vars, size_t number_var);
 ErrorCode clean_stdin       ();
 
 #endif // INPUT_H
