+++ Variables.h
@@ -5,19 +5,19 @@
 
 const char BANNED_SYMBOLS[] = "!@#$%&;:<>()[]{}";
 
-static const int MAX_SIZE_VAR = 100;
+static const size_t MAX_SIZE_VAR = 100;
 struct Variable {
     char* name;
     int valid;
     double value;
 };
 
-static const int DEFAULT_CAPACITY =     1;
-static const int MULTIPLIER_CAPACITY =  2;
+static const size_t DEFAULT_CAPACITY =     1;
+static const size_t MULTIPLIER_CAPACITY =  2;
 struct Variables {
     Variable* var;
-    int count;
-    int capacity;
+    size_t count;
+    size_t capacity;
 };
 
 ErrorCode variables_copy(Variables* vars, Variables** new_vars);
